refactored the unit tests for intro page
